scripts/requirements.txt
# OpenRole Lead Generation Scripts
# ================================

# Core
requests>=2.28.0
python-dotenv>=1.0.0

# PDF Generation
weasyprint>=62.0
Jinja2>=3.0

# Data Processing
pandas>=2.0.0

# Legacy (audit_brand.py)
duckduckgo-search
openai
fake-useragent
